requirements.txt
black
django==2.2.1
psycopg2-binary==2.8.2
djangorestframework
gunicorn==19.9.0
Pillow
arrow
django-phonenumber-field
phonenumbers
drf-yasg
drf-yasg[validation]
moviepy
django-nested-inline
drf-firebase-auth
django-storages
sentry-sdk==0.14.4
fcm-django